requirements.txt
pandas~=2.2.2
scikit-learn~=1.5.1
Flask==3.0.3
torch~=2.4.0+cu124
torchvision~=0.19.0+cu124
grad-cam==1.5.4
pillow~=10.2.0
requests~=2.32.3
PyQt5==5.15.11